clean up lns_addition.cpp and replace long long with int64_t

+++ xlnstorch/xlnstorch/_C/lns_constants.h
@@ -1,6 +1,8 @@
 #ifndef LNS_CONSTANTS_H
 #define LNS_CONSTANTS_H
 
+#include <cstdint>
+
 namespace lns {
 
     /*
@@ -9,13 +11,13 @@ namespace lns {
     different bases.
     */
 
-    inline constexpr long long zero_int = (-(1LL << 53)) | 1LL;
+    inline constexpr int64_t zero_int = (-(1LL << 53)) | 1LL;
     inline constexpr double zero = static_cast<double>(zero_int);
 
-    inline constexpr long long one_int = 0LL;
+    inline constexpr int64_t one_int = 0LL;
     inline constexpr double one = static_cast<double>(one_int);
 
-    inline constexpr long long neg_one_int = 1LL;
+    inline constexpr int64_t neg_one_int = 1LL;
     inline constexpr double neg_one = static_cast<double>(neg_one_int);
 
 }
